this is version 2.0.0
This is my Notty game improvement code, this update adds a difficulty button, and adds the ability for players to throw cards and draw computer player cards, where when the computer player is 2 players, click the draw button to select the player to draw. At the same time, I optimized the number of piles and the hand so that I didn't have more than 80 cards in my hand. I've optimized the game's interface and added a difficulty button, but the open-ended difficulty settings for PC players are not yet complete, and the initial difficulty settings are three levels, Easy, Hard, and Hell. It may be updated later to be more difficult, depending on the design of the code logic. The current time is November 24, 2024, 21:35 in the school library.